# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth Info Service API Documentation",
        "title": "Auth Info API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/dict/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据字典类型编码获取字典数据，按 sort 正序返回。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "获取字典数据列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典类型编码",
                        "name": "type_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.ListDictItemsReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的字典数据项，默认启用。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "创建字典数据",
                "parameters": [
                    {
                        "description": "创建字典数据参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.CreateDictItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
        "/dict/items/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据路径 ID 更新字典数据的键、值、描述、排序和状态。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "更新字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字典数据参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.UpdateDictItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "字典数据不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据路径 ID 软删除字典数据。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "删除字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "字典数据不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
        "/dict/types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取全部字典类型，按 sort 正序返回。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "获取字典类型列表",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.ListDictTypesReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的字典类型，code 必须唯一。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "创建字典类型",
                "parameters": [
                    {
                        "description": "创建字典类型参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.CreateDictTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "409": {
                        "description": "字典类型编码已存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
        "/dict/types/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据路径 ID 更新字典类型名称、描述和排序，code 不可修改。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "更新字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新字典类型参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apipb.UpdateDictTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "字典类型不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据路径 ID 软删除字典类型。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dict"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "403": {
                        "description": "无访问权限",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "404": {
                        "description": "字典类型不存在",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/apipb.OperationReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apipb.CreateDictItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "item_key": {
                    "type": "string"
                },
                "item_value": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "type_code": {
                    "type": "string"
                }
            }
        },
        "apipb.CreateDictTypeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
        "apipb.DictItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_key": {
                    "type": "string"
                },
                "item_value": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type_code": {
                    "type": "string"
                }
            }
        },
        "apipb.DictType": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
        "apipb.ListDictItemsReply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apipb.DictItem"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "apipb.ListDictTypesReply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apipb.DictType"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "apipb.OperationReply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "apipb.UpdateDictItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_key": {
                    "type": "string"
                },
                "item_value": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "apipb.UpdateDictTypeRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}